requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.64.0
google-cloud-storage==2.18.2
langchain==0.2.15
langchain-community==0.2.14
langchain-chroma==0.1.3
fastembed==0.3.6
chromadb==0.5.3
pypdf==4.3.1
streamlit-browser-session-storage==0.0.11
openai==1.51.0